Create interface and default implementation for ManagedObjectRegistry
