+++ platforms/core-configuration/model-core/src/test/groovy/org/gradle/internal/instantiation/managed/ManagedObjectRegistryTest.groovy
@@ -233,7 +233,7 @@ class ManagedObjectRegistryTest extends Specification {
         def parentRegistry = registryOf {
             it.add(ManyProvidingImplementation)
         }
-        def registry = new ManagedObjectRegistry(parentRegistry)
+        def registry = new DefaultManagedObjectRegistry(parentRegistry)
 
         when:
         Thing thing = registry.newInstance(Thing)
@@ -502,7 +502,7 @@ class ManagedObjectRegistryTest extends Specification {
             it.addProvider(new ServiceRegistrationProvider() {
                 @Provides
                 ManagedObjectRegistry createManagedObjectRegistry() {
-                    return new ManagedObjectRegistry(null)
+                    return new DefaultManagedObjectRegistry(null)
                 }
             })
             action.registerServices(it)
